requirements.txt
-i http://mirrors.aliyun.com/pypi/simple/ --trusted-host mirrors.aliyun.com
gitdb===4.0.4
gitpython==3.1.1
smmap==3.0.2